Make `--export` only control .h file generation
It seems that if a programmer uses the `export` keyword then they want
to generate the wrapper functions, so the presence of `export` is the
only thing that should control whether or not the wrapper functions
are generated. On the other hand, the programmer may not always want
to generate the .h file, so the `--export` keyword can control that.